* Run performance check for PBKDF2 from LUKS code, do not mix hash algoritms results. * Add ability to provide pre-generated master key and UUID in LUKS header format. * Add LUKS function to verify master key digest. * Move key slot manuipulation function into LUKS specific code.

git-svn-id: https://cryptsetup.googlecode.com/svn/trunk@94 36d66b0a-2a48-0410-832c-cd162a569da5

+++ luks/pbkdf.h
@@ -11,7 +11,7 @@ int PBKDF2_HMAC(const char *hash,
 		char *dKey, size_t dKeyLen);
 
 
-int PBKDF2_performance_check(const char *hash, unsigned int *iter);
+int PBKDF2_performance_check(const char *hash, uint64_t *iter);
 int PBKDF2_HMAC_ready(const char *hash);
 
 #endif
